fix: linted project and improve files
