+++ server/src/routes/utils.ts
@@ -2,6 +2,7 @@ import multer from "multer";
 import { v4 as uuidv4 } from "uuid";
 import { IMAGES_DIR } from "../common/imageHandler";
 import path from "path";
+import swaggerJSDoc from "swagger-jsdoc";
 
 const storage = multer.diskStorage({
   destination: function (req, file, cb) {
@@ -13,3 +14,16 @@ const storage = multer.diskStorage({
 });
 
 export const uploadImage = multer({ storage });
+
+const options = {
+  definition: {
+    openapi: "3.0.0",
+    info: {
+      title: "MadaMal API",
+      version: "1.0.0",
+    },
+  },
+  apis: [`${__dirname}/routes/*.ts`, `${__dirname}/routes/*.js`], // Path to the API docs
+};
+export const openapiSpecification = swaggerJSDoc(options);
+
